Switch to using resin/* base image.
It seems that current balenalib/* images have a bug that prevents
cross-building the image. The bug is caused by an ONBUILD statement that
calls `cat` binary from the image in context of host which could have a
non-ARM arch.

This patch switches all Dockerfiles based on balenalib/* images back to
resin/* images.

Signed-off-by: Tomasz Kuzemko <tomasz@kuzemko.net>